Build up class hierarchy for config items.
